Moved deep-offset pagination tests out of the legacy suite
ref INC-311

- test/legacy is the old regression suite that's slated for rewrite or
  deletion (see #23611), so new coverage shouldn't be added to it
- the tests are standalone (direct knex seeding, no shared fixtures), so
  they now live in test/integration/models with a single beforeAll seed
  instead of reseeding 1300 rows before every test
